+++ components/blogs/ListBlogs.vue
@@ -24,7 +24,7 @@
 
         <div class="py-8 flex flex-wrap" :class="nuxtSections ? '' : 'md:pl-16'">
           <div v-for="(blog, idx) in blogsResponse" :key="`blog--${blog.id}`" class="m-2">
-            <Card :can-delete="(blogsUserRoleProp.includes('admin') || blogsUserRoleProp.includes('author'))" :description="blog.description" :edit-label="blogsUserRoleProp.includes('publisher') ? blog.published ? $t(mediaTranslationPrefix + 'blogs.unpublish') : $t(mediaTranslationPrefix + 'blogs.publish') : $t(mediaTranslationPrefix + 'blogs.editContent')" :published="blog.published" :draft-of="blog.draft_of ? $t(mediaTranslationPrefix + 'blogs.draftOf', {id: blog.draft_of}) : ''" :media-src="blog.medias && blog.medias[0] && blog.medias[0].files ? blog.medias[0].files[0].url : ''" :blog-title="blog.title && blog.title !== '' && blog.title !== 'null' ? blog.title : blog.medias && blog.medias[0] && blog.medias[0].files ? blog.medias[0].files[0].filename : ''" :blog-title-style="'w-200px overflow-hidden text-ellipsis white whitespace-nowrap'" :blog-author="getAuthorName(blog.author_id)" :is-author="blog.author_id === blogsUserId" :last-update-date="blog.updated ? $t(mediaTranslationPrefix + 'blogs.lastUpdateDate') + parseDate(blog.updated) : ''" :open-blog="() => {blogsUserRoleProp.includes('publisher') ? null : openBlog(blog.id, blog.author_id, blog.default_locale)}" :edit-blog="() => {blogsUserRoleProp.includes('publisher') ? publishBlogByID(blog.id, blog.published, idx) : openBlog(blog.id, blog.author_id, blog.default_locale)}" @delete-blog="blogId = blog.id; showPopup = true" />
+            <Card :can-delete="(blogsUserRoleProp.includes('admin') || blogsUserRoleProp.includes('author'))" :description="blog.description" :edit-label="blogsUserRoleProp.includes('publisher') ? blog.published ? $t(mediaTranslationPrefix + 'blogs.unpublish') : $t(mediaTranslationPrefix + 'blogs.publish') : $t(mediaTranslationPrefix + 'blogs.editContent')" :published="blog.published" :draft-of="blog.draft_of ? $t(mediaTranslationPrefix + 'blogs.draftOf', {id: blog.draft_of}) : ''" :media-src="blog.medias && blog.medias[0] && blog.medias[0].files ? blog.medias[0].files[0].url : ''" :blog-title="blog.title && blog.title !== '' && blog.title !== 'null' ? blog.title : blog.medias && blog.medias[0] && blog.medias[0].files ? blog.medias[0].files[0].filename : ''" :blog-title-style="'w-200px overflow-hidden text-ellipsis white whitespace-nowrap'" :blog-author="getAuthorName(blog.author_id)" :is-author="blog.author_id === blogsUserId" :last-update-date="blog.updated ? $t(mediaTranslationPrefix + 'blogs.lastUpdateDate') + parseDate(blog.updated) : ''" :open-blog="() => {blogsUserRoleProp.includes('publisher') ? null : openBlog(blog.id, blog.author_id, blog.default_locale)}" :edit-blog="() => {blogsUserRoleProp.includes('publisher') ? publishBlogByID(blog.id, blog.published, idx) : openBlog(blog.id, blog.author_id, blog.default_locale)}" @delete-blog="blogId = blog.id; showPopup = true" @open-original-blog="openBlog(blog.draft_of, blog.author_id, blog.default_locale)" />
           </div>
         </div>
 
@@ -552,7 +552,9 @@ export default {
         )
       })
 
-      this.blogsResponse = response.data.data
+      if (response.data && response.data.data && response.data.data.length > 0) {
+        this.blogsResponse = response.data.data.filter(article => !response.data.data.some(atcl => atcl.draft_of === article.id))
+      }
       this.totalBlogs = response.data.total
 
       this.$nuxt.$emit('setModel')