+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - [x] Fixed Blogs components for when a user has multiple roles: When a user has both roles Author and Publisher, or Admin and Publisher, it should have access for the 2 roles eweev/geeks/bo-blogs#22
 
+- [x] Draft management improvement: Remove duplicate articles and show only the draft with a link to access the original article eweev/geeks/bo-blogs#36
 
 ### v1.0.11 (07/01/2025)